Simplify `BindingsBackedCodec` usage
